BUG Error in alias analysis
