+++ crates/roc_host/src/lib.rs
@@ -1,5 +1,6 @@
 mod http_client;
 mod http_server;
+mod json_web_token;
 mod roc;
 mod roc_http;
