+++ crates/roc_host/src/roc.rs
@@ -1,12 +1,9 @@
 #![allow(non_snake_case)]
-use hmac::{Hmac, Mac};
-use jwt::algorithm::AlgorithmType;
+use crate::json_web_token::{JWTFromRoc, JWTToRoc};
 use roc_fn::roc_fn;
 use roc_std::{RocBox, RocList, RocResult, RocStr};
-use sha2::Sha256;
 use std::alloc::Layout;
 use std::cell::RefCell;
-use std::collections::BTreeMap;
 use std::io::{BufRead, BufReader, ErrorKind, Read, Write};
 use std::iter::FromIterator;
 use std::mem::{ManuallyDrop, MaybeUninit};
@@ -949,123 +946,9 @@ pub extern "C" fn roc_fx_tempDir() -> RocResult<RocList<u8>, ()> {
     RocResult::ok(RocList::from(path_os_string_bytes.as_slice()))
 }
 
-#[derive(Clone, Default, Debug, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[repr(C)]
-pub struct JWTFromRoc {
-    secret: RocStr,
-    token: RocStr,
-    algo: u8,
-}
-
-impl JWTFromRoc {
-    fn algorithm(&self) -> AlgorithmType {
-        match self.algo {
-            1 => AlgorithmType::Hs256,
-            2 => AlgorithmType::Hs384,
-            3 => AlgorithmType::Hs512,
-            4 => AlgorithmType::Rs256,
-            5 => AlgorithmType::Rs384,
-            6 => AlgorithmType::Rs512,
-            7 => AlgorithmType::Es256,
-            8 => AlgorithmType::Es384,
-            9 => AlgorithmType::Es512,
-            10 => AlgorithmType::Ps256,
-            11 => AlgorithmType::Ps384,
-            12 => AlgorithmType::Ps512,
-            13 => AlgorithmType::None,
-            _ => panic!("invalid algorithm from roc"),
-        }
-    }
-}
-
-impl roc_std::RocRefcounted for JWTFromRoc {
-    fn inc(&mut self) {
-        self.secret.inc();
-        self.token.inc();
-    }
-    fn dec(&mut self) {
-        self.secret.dec();
-        self.token.dec();
-    }
-    fn is_refcounted() -> bool {
-        true
-    }
-}
-
-#[derive(Clone, Default, Debug, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[repr(C)]
-pub struct JWTToRoc {
-    name: RocStr,
-    value: RocStr,
-}
-
-impl roc_std::RocRefcounted for JWTToRoc {
-    fn inc(&mut self) {
-        self.name.inc();
-        self.value.inc();
-    }
-    fn dec(&mut self) {
-        self.name.dec();
-        self.value.dec();
-    }
-    fn is_refcounted() -> bool {
-        true
-    }
-}
-
 #[roc_fn(name = "jwtVerify")]
 pub extern "C" fn jwt_verify(jwt: &JWTFromRoc) -> RocResult<RocList<JWTToRoc>, RocStr> {
-    use jwt::VerifyWithKey;
-
-    match jwt.algorithm() {
-        AlgorithmType::Hs256 => match h256(&jwt.secret) {
-            Ok(key) => {
-                let token_str = jwt.token.as_str().to_string();
-                let claims: Result<BTreeMap<String, serde_json::Value>, jwt::error::Error> =
-                    token_str.verify_with_key(&key);
-
-                match claims {
-                    Ok(claims) => RocResult::ok(jwt_claims_to_roc(claims)),
-                    Err(err) => RocResult::err(jwt_error_to_roc(err)),
-                }
-            }
-            Err(msg) => RocResult::err(msg),
-        },
-        _ => todo!(),
-    }
-}
-
-fn jwt_claims_to_roc(claims: BTreeMap<String, serde_json::Value>) -> RocList<JWTToRoc> {
-    let mut roc_claims = RocList::with_capacity(claims.len());
-
-    for (name, value) in claims {
-        roc_claims.push(JWTToRoc {
-            name: RocStr::from(name.as_str()),
-            value: RocStr::from(format!("{}", value).to_string().as_str()),
-        });
-    }
-
-    roc_claims
-}
-
-fn jwt_error_to_roc(err: jwt::error::Error) -> RocStr {
-    match err {
-        jwt::error::Error::AlgorithmMismatch(..) => RocStr::from("AlgorithmMismatch"),
-        jwt::error::Error::InvalidSignature => RocStr::from("InvalidSignature"),
-        jwt::error::Error::Format => RocStr::from("BadFormat"),
-        jwt::error::Error::Json(..) => RocStr::from("BadJson"),
-        jwt::error::Error::NoClaimsComponent => RocStr::from("MissingClaims"),
-        jwt::error::Error::NoHeaderComponent => RocStr::from("MissingHeader"),
-        jwt::error::Error::NoKeyId => RocStr::from("MissingKeyId"),
-        _ => RocStr::from(format!("{err}").to_string().as_str()),
-    }
-}
-
-fn h256(str: &RocStr) -> Result<Hmac<Sha256>, RocStr> {
-    match Hmac::new_from_slice(str.as_bytes()) {
-        Ok(key) => Ok(key),
-        Err(err) => Err(RocStr::from(format!("{err}").as_str())),
-    }
+    crate::json_web_token::jwt_verify(jwt)
 }
 
 #[derive(Debug)]
